move more things into external
